--- MyRestWS/.settings/org.eclipse.wst.common.project.facet.core.prefs.xml
@@ -1,7 +1,0 @@
-<root>
-  <facet id="jst.jaxrs">
-    <node name="libprov">
-      <attribute name="provider-id" value="jaxrs-no-op-library-provider"/>
-    </node>
-  </facet>
-</root>
