--- MyRestWS/.settings/org.eclipse.wst.validation.prefs
@@ -1,2 +1,0 @@
-disabled=06target
-eclipse.preferences.version=1
