--- MyRestWS/src/com/cwctravel/rest/api/ExcelService.java
@@ -1,29 +1,0 @@
-package com.cwctravel.rest.api;
-
-import java.io.File;
-
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.core.Response.ResponseBuilder;
-
-@Path("/excel")
-public class ExcelService {
-
-	private static final String FILE_PATH = "c:\\excel-file.xls";
-
-	@GET
-	@Path("/get")
-	@Produces("application/vnd.ms-excel")
-	public Response getFile() {
-
-		File file = new File(FILE_PATH);
-
-		ResponseBuilder response = Response.ok(file);
-		response.header("Content-Disposition", "attachment; filename=new-excel-file.xls");
-		return response.build();
-
-	}
-
-}
