+++ src/com/cwctravel/rest/api/info/Product.java
@@ -1,0 +1,31 @@
+package com.cwctravel.rest.api.info;
+
+import javax.xml.bind.annotation.XmlRootElement;
+
+@XmlRootElement(name = "product")
+public class Product {
+
+	String name;
+	int qty;
+
+	public String getName() {
+		return name;
+	}
+
+	public void setName(String name) {
+		this.name = name;
+	}
+
+	public int getQty() {
+		return qty;
+	}
+
+	public void setQty(int qty) {
+		this.qty = qty;
+	}
+
+	@Override
+	public String toString() {
+		return "Product [name=" + name + ", qty=" + qty + "]";
+	}
+}
